Possible fix Lockup when using AWE32/SB (issue#80)
I have same bug described in issue#80 - when music=AWE32 and sound=SB then fastdoom lockup at the main screen. This fix help for me.

+++ FASTDOOM/ns_awe32.c
@@ -11,9 +11,9 @@
 #define MPU_RESET_CMD 0xff
 #define MPU_ENTER_UART 0x3f
 
-static WORD wSBCBaseAddx; /* Sound Blaster base address */
-static WORD wEMUBaseAddx; /* EMU8000 subsystem base address */
-static WORD wMpuBaseAddx; /* MPU401 base address */
+static int wSBCBaseAddx; /* Sound Blaster base address */
+static int wEMUBaseAddx; /* EMU8000 subsystem base address */
+static int wMpuBaseAddx; /* MPU401 base address */
 
 static unsigned short NoteFlags[128];
